server/requirements.txt
# Servidor
fastapi==0.115.6
uvicorn[standard]==0.32.1

# MongoDB (4.15+ para Python 3.12+)
pymongo>=4.15.2,<4.17

# Excel
openpyxl>=3.1

# Segurança
python-dotenv==1.0.1
bcrypt>=4.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Validação (Pydantic v1 = pure Python, sem Rust; compatível com Python 3.14)
pydantic>=1.10,<2
